requirements.txt
pillow==6.1
torch
tensorflow-gpu==1.13.1
ipython[all]
tqdm
sox
ruamel.yaml
jupyterlab
tqdm
boto3
json5==0.8.5
jsonpickle==1.3
jsonschema==3.2.0
requests
Flask==1.1.1
future==0.17.1
six
ipdb
h5py
html2text
nltk
progressbar
matplotlib
wget 
tensorboardX
pandas
onnx
wget
num2words
librosa
inflect
kaldi-io
marshmallow
unidecode
sentencepiece
snips-nlu==0.20.2
snips-nlu-en==0.2.3
snips-nlu-parsers==0.4.3
snips-nlu-utils==0.9.1
boto3
matplotlib
h5py
youtokentome
Werkzeug==0.16.0
